App.js and Results.js separated
